move url template env to config
