+++ test-apps/display-performance-test-app/README.md
@@ -38,6 +38,10 @@ Performance tests on iOS have more restrictions than performance tests run on ot
 
 ## Environment Variables
 
+* IMJS_NO_DEV_TOOLS
+  * If defined, do not open the electron dev tools on startup
+* IMJS_NO_MAXIMIZE_WINDOW
+  * If defined, don't maximize the electron window on startup
 * IMJS_MAPBOX_KEY
   * If defined, sets the MapBox key for the `MapLayerOptions` as an "access_token".
 * IMJS_BING_MAPS_KEY
@@ -46,12 +50,6 @@ Performance tests on iOS have more restrictions than performance tests run on ot
   * If defined, the API key supplying access to Cesium ION assets.
 * BRIEFCASE_CACHE_LOCATION
   * iModelHost.cacheDir location
-* IMJS_FRONTEND_TILES_URL_TEMPLATE
-  * If defined, specifies the url for @itwin/frontend-tiles to obtain tile trees for spatial views, served over localhost.
-  * The string can include special tokens: {iModel.key} or {iModel.filename}.
-    * e.g.: http://localhost:8080{iModel.key}-tiles/3dft/ or http://localhost:8080/MshX/{iModel.filename}/
-  * These will get replaced by the value of iModel.key or just the filename of that (no path or extension), correspondingly.
-  * Note that the contents of iModel.key will be different on different OSes.
 * Authentication
   * [Headless](https://github.com/iTwin/auth-clients/tree/main/packages/oidc-signin-tool)
     * IMJS_OIDC_HEADLESS (any value, as long as it's set)
@@ -89,7 +87,7 @@ The default configuration file allows you to specify the following:
 You can specify filename options that you wish to ignore:
 When you chose to produce images through an 'image' or 'both' test, the program will name them in the following format: modelName_viewName_renderMode_viewFlagOpts_renderModeOpts_tilePropsOpts.png (for example, SmallTextTest008_Wireframe_-fll+scL+cmL+slL-clp+con_+solShd_+inst.png). If you chose to do a 'readPixels' test, the program will save the images produced from this in the following format: depth/elemId/type_readPixelsSelectorOpts_modelName_viewName_renderMode_viewFlagOpts_renderModeOpts_tilePropsOpts.png (for example, type_+geom+dist_TimingTest01_V0_HiddenLine_-fll+scL+cmL+slL-clp+con_+solShd_+inst.png). The read pixels selector options, view flag options, render mode options, and tile properties options will all be an abbreviated string representation of the actual property. For example, the render mode option 'enableInstancing' would be shown as '+inst' in the filename. The 'filenameOptsToIgnore' property allows you to specify either an array of strings or a single space delimited string, where each string item is an abbreviated string representation of a read pixels selector option, view flag option, render mode option, or tile properties option. These options will not be included in the image names.
 
-You can specify which saved view you wish to use for each test with the viewName property.  If external saved views exist for the local file you can specify which external saved view to use with the extViewName property (you can also specify external saved views by name using the viewName property if the name does not clash with a normal saved view).  You can also specify a saved view to use via the viewString property.  The viewString property must include the _viewname and_viewStatePropsString properties (and optionally the _selectedElements and/or_overrideElements properties) which you can copy from the external saved view file.
+You can specify which saved view you wish to use for each test with the viewName property.  If external saved views exist for the local file you can specify which external saved view to use with the extViewName property (you can also specify external saved views by name using the viewName property if the name does not clash with a normal saved view).  You can also specify a saved view to use via the viewString property.  The viewString property must include the _viewname_ and _viewStatePropsString_ properties (and optionally the _selectedElements_ and/or _overrideElements_ properties) which you can copy from the external saved view file.
 
 If you wish to specify multiple iModels in the same folder, you can use a asterisk wildcard instead of a specific iModel name to test multiple iModels matching the given wildcard. This is NOT case-sensitive. For example, to test all iModels, set modelName to "*". To test all iModels with the word 'Edge' in them, set modelName to "*edge*". To test iModels starting with the word edge, use "edge*". If no modelName is specified, it will default to testing all iModels in the given iModelLocation. This currently does NOT work when using an iModelHub location with iModelHubProject instead of a local directory with iModelLocation.
 
@@ -157,6 +155,10 @@ If given the option of using a local file path or using iModelHub, the program w
 
 * To specify a local file path to use for accessing an imodel, set the "iModelLocation" setting in the json configuration file (ex. "iModelLocation": "D:/models/").
 * To specify a remote iModel, set `iTwinId` and `iModelId`. External saved views will be downloaded as well.
+* To specify the url for @itwin/frontend-tiles to obtain tile trees for spatial views, served over localhost, set "frontendTilesUrlTemplate".  It can include special tokens: {iModel.key} or
+{iModel.filename} (e.g.: "<http://localhost:8080{iModel.key}-tiles/3dft/>" or "<http://localhost:8080/MshX/{iModel.filename}/>"). These will get replaced by the value of iModel.key or just the filename of
+ that (no path or extension), correspondingly.
+ Note that the contents of iModel.key in DPTA is a GUID.
 
 The json config file allows you to specify settings for the entire test run, for a specific model, and for a specific test performed on a given model. Priority for settings will be given first to those for a specific test, then for a specific model, and finally for the entire test run. For example: if transparency is set to true for the entire test run, but a specific test changes transparency to false, that specific test will NOT have transparency even though the rest of the tests run WILL have transparency.
